+++ spires_inversion/spires.h
@@ -3,34 +3,41 @@
 
 // Function declaration for interpolation
 
-double interpolate_idx(double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes, 
+// The big arrays — imagery spectra (spectrum_background / spectrum_target) and
+// the LUT (lut_reflectances) — are stored as float32 to halve memory; every
+// value is promoted to double at read time inside the kernel, so the math and
+// NLopt run in full double precision. Coordinate axes, shade, and the parameter
+// vector stay double. float32 is a property of the storage signature, not a
+// name suffix.
+
+double interpolate_idx(float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
                        int band_idx, double solar_angle_idx, double dust_concentration_idx, double grain_size_idx);
 
 
-double interpolate(double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
+double interpolate(float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
                    double* lut_bands, int len_lut_bands,
                    double* lut_solar_angles, int len_lut_solar_angles,
                    double* lut_dust_concentrations, int len_lut_dust_concentrations,
                    double* lut_grain_sizes, int len_lut_grain_sizes,
                    int band,
-                   double solar_angle, 
-                   double dust_concentration, 
+                   double solar_angle,
+                   double dust_concentration,
                    double grain_size);
 
-std::vector<double> interpolate_all_idx(double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
+std::vector<double> interpolate_all_idx(float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
                                         double solar_angle, double dust, double grain_size);
 
-std::vector<double> interpolate_all(double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes, 
+std::vector<double> interpolate_all(float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
                                     double* lut_bands, int len_lut_bands,
                                     double* lut_solar_angles, int len_lut_solar_angles,
                                     double* lut_dust_concentrations, int len_lut_dust_concentrations,
                                     double* lut_grain_sizes, int len_lut_grain_sizes,
-                                    double solar_angle, 
-                                    double dust_size, 
+                                    double solar_angle,
+                                    double dust_size,
                                     double grain_size);
 
 
-double* interpolate_all_array(double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
+double* interpolate_all_array(float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
                               double* lut_bands, int len_lut_bands,
                               double* lut_solar_angles, int len_lut_solar_angles,
                               double* lut_dust_concentrations, int len_lut_dust_concentrations,
@@ -40,50 +47,50 @@ double* interpolate_all_array(double* lut_reflectances, int n_lut_bands, int n_l
                               double grain_size);
 
 double spectrum_difference(const std::vector<double>& x,
-                           double* spectrum_background, int len_background, 
-                           double* spectrum_target, int len_target,
+                           float* spectrum_background, int len_background,
+                           float* spectrum_target, int len_target,
                            double* spectrum_shade, int len_shade,
-                           double solar_angle,       
+                           double solar_angle,
                            double* lut_bands, int len_lut_bands,
                            double* lut_solar_angles, int len_lut_solar_angles,
                            double* lut_dust_concentrations, int len_lut_dust_concentrations,
-                           double* lut_grain_sizes, int len_lut_grain_sizes,                     
-                           double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
+                           double* lut_grain_sizes, int len_lut_grain_sizes,
+                           float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
 
 double spectrum_difference_scaled(const std::vector<double> &x,
-                                  double* spectrum_background, int len_background,
-                                  double* spectrum_target, int len_target,
+                                  float* spectrum_background, int len_background,
+                                  float* spectrum_target, int len_target,
                                   double* spectrum_shade, int len_shade,
                                   double solar_angle,
                                   double* lut_bands, int len_lut_bands,
                                   double* lut_solar_angles, int len_lut_solar_angles,
                                   double* lut_dust_concentrations, int len_lut_dust_concentrations,
                                   double* lut_grain_sizes, int len_lut_grain_sizes,
-                                  double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
+                                  float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
 
 
 double spectrum_difference_softmax(const std::vector<double>& z,
-                                   double* spectrum_background, int len_background,
-                                   double* spectrum_target, int len_target,
+                                   float* spectrum_background, int len_background,
+                                   float* spectrum_target, int len_target,
                                    double* spectrum_shade, int len_shade,
                                    double solar_angle,
                                    double* lut_bands, int len_lut_bands,
                                    double* lut_solar_angles, int len_lut_solar_angles,
                                    double* lut_dust_concentrations, int len_lut_dust_concentrations,
                                    double* lut_grain_sizes, int len_lut_grain_sizes,
-                                   double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
+                                   float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
 
 
 double spectrum_difference_hybrid(const std::vector<double>& y,
-                                  double* spectrum_background, int len_background,
-                                  double* spectrum_target, int len_target,
+                                  float* spectrum_background, int len_background,
+                                  float* spectrum_target, int len_target,
                                   double* spectrum_shade, int len_shade,
                                   double solar_angle,
                                   double* lut_bands, int len_lut_bands,
                                   double* lut_solar_angles, int len_lut_solar_angles,
                                   double* lut_dust_concentrations, int len_lut_dust_concentrations,
                                   double* lut_grain_sizes, int len_lut_grain_sizes,
-                                  double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
+                                  float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes);
 
 
 std::vector<double> z_to_x(const std::vector<double>& z,
@@ -100,15 +107,15 @@ std::vector<double> y_to_x_hybrid(const std::vector<double>& y);
 std::vector<double> x_to_y_hybrid(const std::vector<double>& x);
 
 
-std::vector<double>  invert(double* spectrum_background, int len_background,
-                           double* spectrum_target, int len_target,
+std::vector<double>  invert(float* spectrum_background, int len_background,
+                           float* spectrum_target, int len_target,
                            double* spectrum_shade, int len_shade,
                            double solar_angle,
                            double* lut_bands, int len_lut_bands,
                            double* lut_solar_angles, int len_lut_solar_angles,
                            double* lut_dust_concentrations, int len_lut_dust_concentrations,
                            double* lut_grain_sizes, int len_lut_grain_sizes,
-                           double* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,                           
+                           float* lut_reflectances, int n_lut_bands, int n_lut_solar_angles, int n_lut_dust_concentrations, int n_lut_grain_sizes,
                            int max_eval,
                            std::vector<double> x0,
                            int algorithm);
